Moving files to facilitate npm package use, adding instructions, changelog
