+++ CHANGELOG.md
@@ -1,0 +1,16 @@
+# Changelog
+All notable changes to this project will be documented in this file.
+
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## [Unreleased]
+- Nothing yet.
+
+## [1.0.0] - 2018-09-11
+
+### Added
+- npm package.
+
+### Changed
+- Server script to uses package export (`src/handler.js`) internally.
