--- src/server.js
@@ -1,4 +1,0 @@
-const http = require('http');
-const handler = require('./handler');
-
-module.exports = config => http.createServer(handler(config));
